refactor(events): split monolithic interactionCreate into modular handlers
- Break down 980+ line file into 9 specialized modules
- Improve maintainability and testability
- Follow SOLID principles with single responsibility per handler
- Reduce complexity while preserving all functionality
